requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
django-filter==24.2
Markdown==3.6
pillow==10.3.0
djangorestframework-simplejwt==5.3.1